countdown component abstraction that fair and vr can use
