+++ src/lib/Components/Countdown/__tests__/CountdownTimer-tests.tsx
@@ -6,13 +6,30 @@ jest.mock("moment", () => {
   return momentMock
 })
 
+import { Flex, Sans, Spacer } from "@artsy/palette"
 // @ts-ignore STRICTNESS_MIGRATION
 import { render } from "enzyme"
 import React from "react"
-import { CountdownTimer } from "../CountdownTimer"
+import { LabeledTicker } from ".."
+import { CountdownProps, CountdownTimer } from "../CountdownTimer"
 
 // @ts-ignore STRICTNESS_MIGRATION
 const dateString = m => new Date(m).toISOString()
+
+const CountdownText: React.SFC<CountdownProps> = ({ duration, label }) =>
+  // @ts-ignore STRICTNESS_MIGRATION
+  label !== "Closed" && (
+    <Flex justifyContent="center" alignItems="center">
+      <LabeledTicker
+        renderSeparator={() => <Spacer mr={0.5} />}
+        textProps={{ color: "white", size: "3t" }}
+        duration={duration}
+      />
+      <Sans size="1" color="white">
+        {label}
+      </Sans>
+    </Flex>
+  )
 
 describe("CountdownTimer", () => {
   beforeEach(() => {
@@ -26,6 +43,7 @@ describe("CountdownTimer", () => {
         formattedOpeningHours="Opens May 10 at 8:22pm"
         startAt="2018-05-10T20:22:42+00:00"
         endAt="2018-05-14T10:24:31+00:00"
+        countdownComponent={CountdownText}
       />
     )
     expect(comp.text()).toContain("Opens May 10 at 8:22pm")
@@ -37,6 +55,7 @@ describe("CountdownTimer", () => {
         formattedOpeningHours="Opens Apr 14 at 8:00pm"
         startAt="2018-04-14T20:00:00+00:00"
         endAt="2018-05-14T20:00:00+00:00"
+        countdownComponent={CountdownText}
       />
     )
     expect(comp.text()).toContain("Opens Apr 14 at 8:00pm")
@@ -48,6 +67,7 @@ describe("CountdownTimer", () => {
         formattedOpeningHours="Closed"
         startAt={dateString(Date.now() - 2000)}
         endAt={dateString(Date.now() - 1000)}
+        countdownComponent={CountdownText}
       />
     )
     expect(comp.text()).toContain("")
